# Add performance test for tenant start