--- a/test_runner/performance/test_bulk_tenant_create.py
+++ b/test_runner/performance/test_bulk_tenant_create.py
@@ -10,12 +10,13 @@ from fixtures.zenith_fixtures import (
 
 pytest_plugins = ("fixtures.benchmark_fixture")
 
-# Run bulk tenant cration test.
+# Run bulk tenant creation test.
 #
 # Collects metrics:
 #
 # 1. Time to create {1,10,50} tenants
 # 2. Average creation time per tenant
+
 
 @pytest.mark.parametrize('tenants_count', [1, 10, 50])
 def test_bulk_tenant_create(
@@ -29,7 +30,9 @@ def test_bulk_tenant_create(
     start = timeit.default_timer()
     for i in range(tenants_count):
         tenant = tenant_factory.create()
-        zenith_cli.run(["branch", f"test_bulk_tenant_create_{tenants_count}_{i}", "main", f"--tenantid={tenant}"])
+        zenith_cli.run([
+            "branch", f"test_bulk_tenant_create_{tenants_count}_{i}", "main", f"--tenantid={tenant}"
+        ])
         pg_tenant = postgres.create_start(
             f"test_bulk_tenant_create_{tenants_count}_{i}",
             tenant,
@@ -38,7 +41,7 @@ def test_bulk_tenant_create(
         with closing(pg_tenant.connect()) as conn:
             with conn.cursor() as cur:
                 cur.execute("SELECT 1;")
-                assert cur.fetchone() == (1,)
+                assert cur.fetchone() == (1, )
     end = timeit.default_timer()
 
     zenbenchmark.record('tenant_creation_time_total', end - start, 's')