test have 3 items: create 1, 10 and 50 tenants

example of test output

```
test_bulk_tenant_create[1].tenant_creation_time_total: 1.677 s
test_bulk_tenant_create[1].tenant_creation_time_avg: 1.677 s
test_bulk_tenant_create[10].tenant_creation_time_total: 17.489 s
test_bulk_tenant_create[10].tenant_creation_time_avg: 1.749 s
test_bulk_tenant_create[50].tenant_creation_time_total: 91.056 s
test_bulk_tenant_create[50].tenant_creation_time_avg: 1.821 s
```