@cicdteam 2021-10-03
test modified:
- variant with wal_acceptors added
- tenant count changed to `[1,5,10]` as postgre has `MAX_WALKEEPERS=32` (and creating 11th tenant with 3 wal_acceptors each broke this limit)
- removed `SELECT 1` query as main goal - measure tenant creation time exactly
- renamed output results, average time printed only (total was removed as it makes no sense)

example of test output:

```
test_bulk_tenant_create[with_wa-1].tenant_creation_time: 4.006 s
test_bulk_tenant_create[with_wa-5].tenant_creation_time: 4.183 s
test_bulk_tenant_create[with_wa-10].tenant_creation_time: 4.571 s
test_bulk_tenant_create[without_wa-1].tenant_creation_time: 4.198 s
test_bulk_tenant_create[without_wa-5].tenant_creation_time: 4.337 s
test_bulk_tenant_create[without_wa-10].tenant_creation_time: 4.043 s
```